#19 Remaster services layer

+++ src/main/java/market/domain/Bill.java
@@ -10,7 +10,7 @@ import java.util.Date;
 import static javax.persistence.TemporalType.TIMESTAMP;
 
 /**
- * Счёт для оплаты заказа.
+ * Bill of the {@link Order}.
  */
 @Entity
 @Table(name = "bill")
@@ -29,7 +29,7 @@ public class Bill implements Serializable {
 	private int number;
 
 	@Column(name = "total_cost", nullable = false)
-	private int totalCost;
+	private double totalCost;
 
 	@Column(name = "payed", nullable = false)
 	private boolean payed = false;
@@ -51,100 +51,58 @@ public class Bill implements Serializable {
 		this.ccNumber = ccNumber;
 	}
 
-	/**
-	 * @return the id
-	 */
 	public Long getId() {
 		return id;
 	}
 
-	/**
-	 * @param id the id to set
-	 */
 	public void setId(Long id) {
 		this.id = id;
 	}
 
-	/**
-	 * @return the ccNumber
-	 */
 	public String getCcNumber() {
 		return ccNumber;
 	}
 
-	/**
-	 * @param ccNumber the ccNumber to set
-	 */
 	public void setCcNumber(String ccNumber) {
 		this.ccNumber = ccNumber;
 	}
 
-	/**
-	 * @return the order
-	 */
 	public Order getOrder() {
 		return order;
 	}
 
-	/**
-	 * @param order the order to set
-	 */
 	public void setOrder(Order order) {
 		this.order = order;
 	}
 
-	/**
-	 * @return the number
-	 */
 	public int getNumber() {
 		return number;
 	}
 
-	/**
-	 * @param number the number to set
-	 */
 	public void setNumber(int number) {
 		this.number = number;
 	}
 
-	/**
-	 * @return the dateCreated
-	 */
 	public Date getDateCreated() {
 		return dateCreated;
 	}
 
-	/**
-	 * @param dateCreated the dateCreated to set
-	 */
 	public void setDateCreated(Date dateCreated) {
 		this.dateCreated = dateCreated;
 	}
 
-	/**
-	 * @return the totalCost
-	 */
-	public int getTotalCost() {
+	public double getTotalCost() {
 		return totalCost;
 	}
 
-	/**
-	 * @param totalCost the totalCost to set
-	 */
-	public void setTotalCost(int totalCost) {
+	public void setTotalCost(double totalCost) {
 		this.totalCost = totalCost;
 	}
 
-	/**
-	 * @return the payed
-	 */
 	public boolean isPayed() {
 		return payed;
 	}
 
-	/**
-	 * @param payed the payed to set
-	 */
 	public void setPayed(boolean payed) {
 		this.payed = payed;
 	}
